Now, only model node can be double clicked to be built
